@@ -1,30 +1,36 @@
-module github.com/fastbill/go-service-toolkit/v3
+module github.com/fastbill/go-service-toolkit/v4
 
-go 1.13
+go 1.15
 
 require (
-	github.com/alicebob/miniredis/v2 v2.10.0
-	github.com/fastbill/go-httperrors/v2 v2.0.0
+	github.com/alicebob/miniredis/v2 v2.14.1
+	github.com/fastbill/go-httperrors/v2 v2.0.1
 	github.com/getsentry/sentry-go v0.2.1
-	github.com/go-playground/universal-translator v0.17.0 // indirect
-	github.com/go-playground/validator v9.21.0+incompatible
-	github.com/go-redis/redis/v7 v7.0.0-beta.4
-	github.com/golang-migrate/migrate/v4 v4.6.2
-	github.com/gomodule/redigo v1.7.1-0.20190724094224-574c33c3df38 // indirect
-	github.com/google/go-cmp v0.4.0 // indirect
-	github.com/jinzhu/gorm v1.9.8
+	github.com/go-playground/validator/v10 v10.4.1
+	github.com/go-redis/redis/v8 v8.4.8
+	github.com/golang-migrate/migrate/v4 v4.14.1
+	github.com/hashicorp/errwrap v1.1.0 // indirect
+	github.com/jackc/pgproto3/v2 v2.0.7 // indirect
 	github.com/joho/godotenv v1.3.0
 	github.com/labstack/echo/v4 v4.1.6
 	github.com/labstack/gommon v0.3.0
-	github.com/onsi/ginkgo v1.10.3 // indirect
-	github.com/onsi/gomega v1.7.1 // indirect
+	github.com/leodido/go-urn v1.2.1 // indirect
+	github.com/lib/pq v1.9.0 // indirect
+	github.com/mattn/go-colorable v0.1.8 // indirect
 	github.com/pingcap/errors v0.11.4 // indirect
-	github.com/pkg/errors v0.8.1
-	github.com/prometheus/client_golang v1.1.0
+	github.com/prometheus/client_golang v1.9.0
 	github.com/sirupsen/logrus v1.7.0
-	github.com/stretchr/testify v1.4.0
-	golang.org/x/crypto v0.0.0-20191227163750-53104e6ec876 // indirect
-	golang.org/x/net v0.0.0-20191209160850-c0dbc17a3553 // indirect
-	golang.org/x/sys v0.0.0-20201202213521-69691e467435 // indirect
-	gopkg.in/go-playground/assert.v1 v1.2.1 // indirect
+	github.com/stretchr/objx v0.3.0 // indirect
+	github.com/stretchr/testify v1.6.1
+	github.com/valyala/fasttemplate v1.2.1 // indirect
+	github.com/yuin/gopher-lua v0.0.0-20200816102855-ee81675732da // indirect
+	golang.org/x/crypto v0.0.0-20201221181555-eec23a3978ad // indirect
+	golang.org/x/net v0.0.0-20201224014010-6772e930b67b // indirect
+	golang.org/x/sys v0.0.0-20210110051926-789bb1bd4061 // indirect
+	golang.org/x/text v0.3.5 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+	gorm.io/driver/mysql v1.0.3
+	gorm.io/driver/postgres v1.0.6
+	gorm.io/gorm v1.20.11
 )
