@@ -2,6 +2,7 @@ package server
 
 import (
 	"context"
+	"fmt"
 	"net/http"
 	"os"
 	"os/signal"
@@ -10,10 +11,9 @@ import (
 	"time"
 
 	"github.com/fastbill/go-httperrors/v2"
-	"github.com/fastbill/go-service-toolkit/v3/observance"
+	"github.com/fastbill/go-service-toolkit/v4/observance"
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
-	"github.com/pkg/errors"
 )
 
 const defaultTimeout = 30 * time.Second
@@ -27,7 +27,7 @@ func New(obs *observance.Obs, CORSOrigins string, timeout ...string) (*echo.Echo
 	if len(timeout) > 0 {
 		parsedTimeout, err := time.ParseDuration(timeout[0])
 		if err != nil {
-			return nil, nil, errors.Wrap(err, "timeout could not be parsed")
+			return nil, nil, fmt.Errorf("timeout could not be parsed: %w", err)
 		}
 		timeoutDuration = parsedTimeout
 	}
